+++ security/base-identities/sso.tf
@@ -1,0 +1,4 @@
+/*resource "aws_iam_saml_provider" "this" {
+  name                   = "${var.project}-${var.environment}-sso-jumpcloud"
+  saml_metadata_document = ""
+}*/