+++ CHANGELOG.md
@@ -7,6 +7,60 @@ All notable changes to this project will be documented in this file.
 
 
 
+<a name="v1.3.68"></a>
+## [v1.3.68] - 2021-10-08
+
+- Create FUNDING.yml
+
+
+<a name="v1.3.67"></a>
+## [v1.3.67] - 2021-10-07
+
+- Refactor all accounts directories to add region subdirectories ([#314](https://github.com/binbashar/bb-devops-tf-infra-aws/issues/314))
+
+
+<a name="v1.3.66"></a>
+## [v1.3.66] - 2021-10-06
+
+- Small sso org enabling change
+- Merge branch 'master' into feature/open-soruce-repo
+- sso + jumpcloud
+
+
+<a name="v1.3.65"></a>
+## [v1.3.65] - 2021-10-06
+
+- Add more permissions to DevOps role in order try out more AWS services ([#312](https://github.com/binbashar/bb-devops-tf-infra-aws/issues/312))
+
+
+<a name="v1.3.64"></a>
+## [v1.3.64] - 2021-10-01
+
+- Change tag refs to lowerCamelCase
+- Change FirewallManager resource_tags to true
+- Fix tfstate path
+- Rename wrong filename
+- Remove AWS config authorization resources
+- Remove AWS config authorization resources
+- Remove AWS config authorization resources
+- Add AWS Config delegation to the Security account
+- Add AWS Config delegation to the Security account
+- Update resource_tags to use FirewallManager tag
+- Add AWS Config agregator into the Security account
+
+
+<a name="v1.3.63"></a>
+## [v1.3.63] - 2021-09-30
+
+- Add resource dependecies for FMS
+- Change Network Firewall module version
+- Update terraform-aws-firewall-manager module source to Binbash release
+- * Update terraform-aws-firewall-manager module source to Binbash release * Update fms policies & rules
+- Add DNS Firewall rules support in FMS
+- Ad Network Firewall rules definitions
+- Add Network Firewall Policies
+
+
 <a name="v1.3.62"></a>
 ## [v1.3.62] - 2021-09-17
 
@@ -2122,7 +2176,13 @@ All notable changes to this project will be documented in this file.
 - BBL-16 binbash org baseline/placeholders files
 
 
-[Unreleased]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.62...HEAD
+[Unreleased]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.68...HEAD
+[v1.3.68]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.67...v1.3.68
+[v1.3.67]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.66...v1.3.67
+[v1.3.66]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.65...v1.3.66
+[v1.3.65]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.64...v1.3.65
+[v1.3.64]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.63...v1.3.64
+[v1.3.63]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.62...v1.3.63
 [v1.3.62]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.61...v1.3.62
 [v1.3.61]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.60...v1.3.61
 [v1.3.60]: https://github.com/binbashar/bb-devops-tf-infra-aws/compare/v1.3.59...v1.3.60